ML/agents/crewai/requirements.txt
crewai>=0.175.0
openai>=1.13.3
python-dotenv>=1.0.0

